Fix subject ready state in stories Use subject viewer actions to disable task stories when the subject is loading. Move MockTask into a shared component.
