+++ packages/lib-classifier/src/helpers/MockTask/index.js
@@ -1,0 +1,1 @@
+export { default } from './MockTask'
