+++ packages/lib-classifier/src/helpers/index.js
@@ -1,6 +1,7 @@
 export { default as createLocationCounts } from './createLocationCounts'
 export { default as findLocationsByMediaType } from './findLocationsByMediaType'
 export { default as layouts } from './layouts'
+export { default as MockTask } from './MockTask'
 export { default as shownMarks } from './shownMarks'
 export { default as subjectViewers } from './subjectViewers'
 export { default as validateMimeType } from './validateMimeType'
