+++ packages/lib-classifier/src/plugins/tasks/DrawingTask/components/DrawingTask.stories.js
@@ -1,11 +1,10 @@
 import { withKnobs, boolean, select } from '@storybook/addon-knobs'
 import asyncStates from '@zooniverse/async-states'
-import { storiesOf } from '@storybook/react'
 import zooTheme from '@zooniverse/grommet-theme'
+import { storiesOf } from '@storybook/react'
 import React from 'react'
-import { Box, Grommet } from 'grommet'
 import { Provider } from 'mobx-react'
-import { Tasks } from '../../../../components/Classifier/components/TaskArea/components/Tasks/Tasks'
+import { MockTask } from '@helpers'
 import { createStore }  from '@store/helpers'
 
 const store = createStore()
@@ -13,34 +12,6 @@ function addStepToStore(step, tasks) {
   const steps = [['S1', step]]
   store.workflowSteps.setStepsAndTasks({ steps, tasks })
   store.workflowSteps.active.tasks.forEach(task => store.classifications.addAnnotation(task))
-}
-
-function MockTask(props) {
-  const { dark, ...taskProps } = props
-
-  return (
-    <Grommet
-      background={{
-        dark: 'dark-1',
-        light: 'light-1'
-      }}
-      theme={Object.assign({}, zooTheme, { dark })}
-      themeMode={(dark) ? 'dark' : 'light'}
-    >
-      <Box
-        background={{
-          dark: 'dark-3',
-          light: 'neutral-6'
-        }}
-        pad='1em'
-        width='380px'
-      >
-        <Tasks
-          {...taskProps}
-        />
-      </Box>
-    </Grommet>
-  )
 }
 
 storiesOf('Tasks / Drawing Task', module)
@@ -106,6 +77,7 @@ storiesOf('Tasks / Drawing Task', module)
           loadingState={asyncStates.success}
           step={store.workflowSteps.active}
           subjectReadyState={subjectReadyState}
+          subjectViewer={store.subjectViewer}
         />
       </Provider>
     )
